+++ CHANGELOG.md
@@ -44,3 +44,9 @@
 ## [1.2.0] - 2026-03-26
 
  - add order endpoint, order model
+
+## [1.2.1] - 2026-03-30
+
+ - add changeSavingPoints to order model
+ - add reduction to order model
+ - add count products to order model